Addressed review comments
PubSub is now in concurrent package
