+++ core/shared/src/main/scala/fs2/concurrent/Queue.scala
@@ -4,8 +4,9 @@ package concurrent
 import cats.{Applicative, Functor}
 import cats.effect.{Concurrent, Sync}
 import cats.implicits._
-import fs2.concurrent.pubsub._
 import fs2.internal.Token
+
+import scala.collection.immutable.{Queue => ScalaQueue}
 
 /** Provides the ability to enqueue elements to a `Queue`. */
 trait Enqueue[F[_], A] {
@@ -111,39 +112,48 @@ object Queue {
 
   /** Creates a queue with no size bound. */
   def unbounded[F[_], A](implicit F: Concurrent[F]): F[Queue[F, A]] =
-    forStrategy(UnboundedQueue.lifo[A])
+    forStrategy(Strategy.lifo[A])
 
   /** unbounded queue that distributed always at max `fairSize` elements to any subscriber **/
   def fairUnbounded[F[_], A](fairSize: Int)(implicit F: Concurrent[F]): F[Queue[F, A]] =
-    forStrategy(UnboundedQueue.lifo[A].transformSelector[Int]((sz, _) => sz.min(fairSize)))
+    forStrategy(Strategy.lifo[A].transformSelector[Int]((sz, _) => sz.min(fairSize)))
 
   /** Creates a queue with the specified size bound. */
   def bounded[F[_], A](maxSize: Int)(implicit F: Concurrent[F]): F[Queue[F, A]] =
-    forStrategy(Bounded.lifo(maxSize))
+    forStrategy(Strategy.boundedLifo(maxSize))
 
-  /** bounded queue that distributed always at max `fairSize` elements to any subscriber **/
-  def fairBounded[F[_], A](maxSize: Int, fairSize: Int)(implicit F: Concurrent[F]): F[Queue[F, A]] =
-    forStrategy(Bounded.lifo(maxSize).transformSelector[Int]((sz, _) => sz.min(fairSize)))
+  /** bounded queue terminated by enqueueing None for the first time. All element before None are preserved in Queue. **/
+  def boundedNoneTerminated[F[_], A](maxSize: Int)(
+      implicit F: Concurrent[F]): F[NoneTerminatedQueue[F, A]] =
+    forStrategyNoneTerminated(PubSub.Strategy.closeDrainFirst(Strategy.boundedLifo(maxSize)))
 
   /** Creates a queue which stores the last `maxSize` enqueued elements and which never blocks on enqueue. */
   def circularBuffer[F[_], A](maxSize: Int)(implicit F: Concurrent[F]): F[Queue[F, A]] =
-    forStrategy(UnboundedQueue.circularBuffer(maxSize))
+    forStrategy(Strategy.circularBuffer(maxSize))
+
+  /** bounded queue that distributed always at max `fairSize` elements to any subscriber **/
+  def fairBounded[F[_], A](maxSize: Int, fairSize: Int)(implicit F: Concurrent[F]): F[Queue[F, A]] =
+    forStrategy(Strategy.boundedLifo(maxSize).transformSelector[Int]((sz, _) => sz.min(fairSize)))
+
+  /** unbounded queue terminated by enqueueing None for the first time. All element before None are preserved in Queue. **/
+  def noneTerminated[F[_], A](implicit F: Concurrent[F]): F[NoneTerminatedQueue[F, A]] =
+    forStrategyNoneTerminated(PubSub.Strategy.closeDrainFirst(Strategy.lifo))
 
   /** Creates a queue which allows a single element to be enqueued at any time. */
   def synchronous[F[_], A](implicit F: Concurrent[F]): F[Queue[F, A]] =
-    forStrategy(Bounded.synchronous)
+    forStrategy(Strategy.synchronous)
 
   /** Like [[synchronous]], except that any enqueue of `None` will never block and cancels any dequeue operation */
   def synchronousNoneTerminated[F[_], A](implicit F: Concurrent[F]): F[NoneTerminatedQueue[F, A]] =
-    forStrategyNoneTerminated(NoneTerminated.closeNow(Bounded.synchronous))
+    forStrategyNoneTerminated(PubSub.Strategy.closeNow(Strategy.synchronous))
 
   private[fs2] def headUnsafe[F[_]: Sync, A](chunk: Chunk[A]): F[A] =
     if (chunk.size == 1) Applicative[F].pure(chunk(0))
     else Sync[F].raiseError(new Throwable(s"Expected chunk of size 1. got $chunk"))
 
   /** creates queue from supplied strategy **/
-  def forStrategy[F[_]: Concurrent, S, A](
-      strategy: PubSubStrategy[A, Chunk[A], S, Int]): F[Queue[F, A]] =
+  private[fs2] def forStrategy[F[_]: Concurrent, S, A](
+      strategy: PubSub.Strategy[A, Chunk[A], S, Int]): F[Queue[F, A]] =
     PubSub(strategy).map { pubSub =>
       new Queue[F, A] {
 
@@ -169,8 +179,9 @@ object Queue {
     }
 
   /** creates queue that is terminated by enqueueing None for supplied strategy **/
-  def forStrategyNoneTerminated[F[_]: Concurrent, S, A](
-      strategy: PubSubStrategy[Option[A], Option[Chunk[A]], S, Int]): F[NoneTerminatedQueue[F, A]] =
+  private[fs2] def forStrategyNoneTerminated[F[_]: Concurrent, S, A](
+      strategy: PubSub.Strategy[Option[A], Option[Chunk[A]], S, Int])
+    : F[NoneTerminatedQueue[F, A]] =
     PubSub(strategy).map { pubSub =>
       new NoneTerminatedQueue[F, A] {
         def enqueue1(a: Option[A]): F[Unit] =
@@ -200,6 +211,97 @@ object Queue {
       }
     }
 
+  private[fs2] object Strategy {
+
+    /** unbounded fifo strategy **/
+    def boundedFifo[A](maxSize: Int): PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] =
+      PubSub.Strategy.bounded(maxSize)(fifo[A])(_.size)
+
+    /** unbounded lifo strategy **/
+    def boundedLifo[A](maxSize: Int): PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] =
+      PubSub.Strategy.bounded(maxSize)(lifo[A])(_.size)
+
+    /** unbounded strategy of CyclicBuffer, that memoize always `maxSize` elements and never blocks. **/
+    def circularBuffer[A](maxSize: Int): PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] =
+      unbounded { (q, a) =>
+        if (q.size < maxSize) q :+ a
+        else q.tail :+ a
+      }
+
+    /** unbounded fifo strategy **/
+    def fifo[A]: PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] = unbounded(_ :+ _)
+
+    /** unbounded lifo strategy **/
+    def lifo[A]: PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] = unbounded((q, a) => a +: q)
+
+    /**
+      * Strategy that allows only single element to be published.
+      * Before the `A` is published, at least one subscriber must be ready to consume.
+      */
+    def synchronous[A]: PubSub.Strategy[A, Chunk[A], (Boolean, Option[A]), Int] =
+      new PubSub.Strategy[A, Chunk[A], (Boolean, Option[A]), Int] {
+        def initial: (Boolean, Option[A]) = (false, None)
+
+        def accepts(i: A, queueState: (Boolean, Option[A])): Boolean =
+          queueState._1 && queueState._2.isEmpty
+
+        def publish(i: A, queueState: (Boolean, Option[A])): (Boolean, Option[A]) =
+          (queueState._1, Some(i))
+
+        def get(selector: Int,
+                queueState: (Boolean, Option[A])): ((Boolean, Option[A]), Option[Chunk[A]]) =
+          queueState._2 match {
+            case None    => ((true, None), None)
+            case Some(a) => ((false, None), Some(Chunk.singleton(a)))
+          }
+
+        def empty(queueState: (Boolean, Option[A])): Boolean =
+          queueState._2.isEmpty
+
+        def subscribe(selector: Int,
+                      queueState: (Boolean, Option[A])): ((Boolean, Option[A]), Boolean) =
+          (queueState, false)
+
+        def unsubscribe(selector: Int, queueState: (Boolean, Option[A])): (Boolean, Option[A]) =
+          queueState
+      }
+
+    /**
+      * Creates unbounded queue strategy for `A` with configurable append function
+      *
+      * @param append Function used to append new elements to the queue.
+      */
+    def unbounded[A](append: (ScalaQueue[A], A) => ScalaQueue[A])
+      : PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] =
+      new PubSub.Strategy[A, Chunk[A], ScalaQueue[A], Int] {
+
+        val initial: ScalaQueue[A] = ScalaQueue.empty
+
+        def publish(a: A, queueState: ScalaQueue[A]): ScalaQueue[A] =
+          append(queueState, a)
+
+        def accepts(i: A, queueState: ScalaQueue[A]): Boolean =
+          true
+
+        def empty(queueState: ScalaQueue[A]): Boolean =
+          queueState.isEmpty
+
+        def get(selector: Int, queueState: ScalaQueue[A]): (ScalaQueue[A], Option[Chunk[A]]) =
+          if (queueState.isEmpty) (queueState, None)
+          else {
+            val (out, rem) = queueState.splitAt(selector)
+            (rem, Some(Chunk.seq(out)))
+          }
+
+        def subscribe(selector: Int, queueState: ScalaQueue[A]): (ScalaQueue[A], Boolean) =
+          (queueState, false)
+
+        def unsubscribe(selector: Int, queueState: ScalaQueue[A]): ScalaQueue[A] =
+          queueState
+      }
+
+  }
+
 }
 
 trait InspectableQueue[F[_], A] extends Queue[F, A] {
@@ -227,24 +329,24 @@ object InspectableQueue {
 
   /** Creates a queue with no size bound. */
   def unbounded[F[_], A](implicit F: Concurrent[F]): F[InspectableQueue[F, A]] =
-    forStrategy(UnboundedQueue.lifo[A])(_.headOption)(_.size)
+    forStrategy(Queue.Strategy.lifo[A])(_.headOption)(_.size)
 
   /** Creates a queue with the specified size bound. */
   def bounded[F[_], A](maxSize: Int)(implicit F: Concurrent[F]): F[InspectableQueue[F, A]] =
-    forStrategy(Bounded.lifo[A](maxSize))(_.headOption)(_.size)
+    forStrategy(Queue.Strategy.boundedLifo[A](maxSize))(_.headOption)(_.size)
 
   /** Creates a queue which stores the last `maxSize` enqueued elements and which never blocks on enqueue. */
   def circularBuffer[F[_], A](maxSize: Int)(implicit F: Concurrent[F]): F[InspectableQueue[F, A]] =
-    forStrategy(UnboundedQueue.circularBuffer[A](maxSize))(_.headOption)(_.size)
+    forStrategy(Queue.Strategy.circularBuffer[A](maxSize))(_.headOption)(_.size)
 
-  def forStrategy[F[_]: Concurrent, S, A](
-      strategy: PubSubStrategy[A, Chunk[A], S, Int]
+  private[fs2] def forStrategy[F[_]: Concurrent, S, A](
+      strategy: PubSub.Strategy[A, Chunk[A], S, Int]
   )(
       headOf: S => Option[A]
   )(
       sizeOf: S => Int
   ): F[InspectableQueue[F, A]] =
-    PubSub(Inspectable.strategy(strategy)).map { pubSub =>
+    PubSub(PubSub.Strategy.Inspectable.strategy(strategy)).map { pubSub =>
       new InspectableQueue[F, A] {
         def enqueue1(a: A): F[Unit] = pubSub.publish(a)
         def offer1(a: A): F[Boolean] = pubSub.tryPublish(a)
