+++ core/shared/src/main/scala/fs2/concurrent/Topic.scala
@@ -3,9 +3,8 @@ package concurrent
 
 import cats.syntax.all._
 import cats.effect.{Concurrent, Sync}
-import scala.collection.immutable.{Queue => ScalaQueue}
 
-import fs2.concurrent.pubsub.{Inspectable, PubSub, TopicStrategy}
+import scala.collection.immutable.{Queue => ScalaQueue}
 import fs2.internal.Token
 
 /**
@@ -85,62 +84,117 @@ abstract class Topic[F[_], A] { self =>
 object Topic {
 
   def apply[F[_], A](initial: A)(implicit F: Concurrent[F]): F[Topic[F, A]] =
-    PubSub(Inspectable.strategy(TopicStrategy.boundedSubscribers(initial))).map { pubSub =>
-      new Topic[F, A] {
+    PubSub(PubSub.Strategy.Inspectable.strategy(Strategy.boundedSubscribers(initial))).map {
+      pubSub =>
+        new Topic[F, A] {
 
-        def subscriber(size: Int): Stream[F, ((Token, Int), Stream[F, ScalaQueue[A]])] =
-          Stream
-            .bracket(
-              Sync[F]
-                .delay((new Token, size))
-                .flatTap(selector => pubSub.subscribe(Right(selector)))
-            )(selector => pubSub.unsubscribe(Right(selector)))
-            .map { selector =>
-              selector ->
-                Stream
-                  .repeatEval(pubSub.get(Right(selector)))
-                  .flatMap {
-                    case Right(q) => Stream.emit(q)
-                    case Left(_)  => Stream.empty // impossible
+          def subscriber(size: Int): Stream[F, ((Token, Int), Stream[F, ScalaQueue[A]])] =
+            Stream
+              .bracket(
+                Sync[F]
+                  .delay((new Token, size))
+                  .flatTap(selector => pubSub.subscribe(Right(selector)))
+              )(selector => pubSub.unsubscribe(Right(selector)))
+              .map { selector =>
+                selector ->
+                  Stream
+                    .repeatEval(pubSub.get(Right(selector)))
+                    .flatMap {
+                      case Right(q) => Stream.emit(q)
+                      case Left(_)  => Stream.empty // impossible
+                    }
+
+              }
+
+          def publish: Sink[F, A] =
+            _.evalMap(publish1)
+
+          def publish1(a: A): F[Unit] =
+            pubSub.publish(a)
+
+          def subscribe(maxQueued: Int): Stream[F, A] =
+            subscriber(maxQueued).flatMap { case (_, s) => s.flatMap(Stream.emits) }
+
+          def subscribeSize(maxQueued: Int): Stream[F, (A, Int)] =
+            subscriber(maxQueued).flatMap {
+              case (selector, stream) =>
+                stream
+                  .flatMap { q =>
+                    Stream.emits(q.zipWithIndex.map { case (a, idx) => (a, q.size - idx) })
                   }
-
+                  .evalMap {
+                    case (a, remQ) =>
+                      pubSub.get(Left(None)).map {
+                        case Left(s) =>
+                          (a, s.subcribers.get(selector).map(_.size + remQ).getOrElse(remQ))
+                        case Right(_) => (a, -1) // impossible
+                      }
+                  }
             }
 
-        def publish: Sink[F, A] =
-          _.evalMap(publish1)
-
-        def publish1(a: A): F[Unit] =
-          pubSub.publish(a)
-
-        def subscribe(maxQueued: Int): Stream[F, A] =
-          subscriber(maxQueued).flatMap { case (_, s) => s.flatMap(Stream.emits) }
-
-        def subscribeSize(maxQueued: Int): Stream[F, (A, Int)] =
-          subscriber(maxQueued).flatMap {
-            case (selector, stream) =>
-              stream
-                .flatMap { q =>
-                  Stream.emits(q.zipWithIndex.map { case (a, idx) => (a, q.size - idx) })
+          def subscribers: Stream[F, Int] =
+            Stream
+              .bracket(Sync[F].delay(new Token))(token => pubSub.unsubscribe(Left(Some(token))))
+              .flatMap { token =>
+                Stream.repeatEval(pubSub.get(Left(Some(token)))).flatMap {
+                  case Left(s)  => Stream.emit(s.subcribers.size)
+                  case Right(_) => Stream.empty //impossible
                 }
-                .evalMap {
-                  case (a, remQ) =>
-                    pubSub.get(Left(None)).map {
-                      case Left(s) =>
-                        (a, s.subcribers.get(selector).map(_.size + remQ).getOrElse(remQ))
-                      case Right(_) => (a, -1) // impossible
-                    }
-                }
+              }
+        }
+    }
+
+  object Strategy {
+
+    case class State[A](
+        last: A,
+        subcribers: Map[(Token, Int), ScalaQueue[A]]
+    )
+
+    /**
+      * Strategy for topic, where every subscriber can specify max size of queued elements.
+      * If that subscription is exceeded any other `publish` to the topic will hold,
+      * until such subscriber disappears, or consumes more elements.
+      *
+      * @param initial  Initial value of the topic.
+      */
+    def boundedSubscribers[F[_], A](
+        start: A): PubSub.Strategy[A, ScalaQueue[A], State[A], (Token, Int)] =
+      new PubSub.Strategy[A, ScalaQueue[A], State[A], (Token, Int)] {
+        def initial: State[A] = State(start, Map.empty)
+        def accepts(i: A, queueState: State[A]): Boolean =
+          queueState.subcribers.forall { case ((_, max), q) => q.size < max }
+
+        def publish(i: A, queueState: State[A]): State[A] =
+          State(
+            last = i,
+            subcribers = queueState.subcribers.mapValues(_ :+ i)
+          )
+
+        def get(selector: (Token, Int), queueState: State[A]): (State[A], Option[ScalaQueue[A]]) =
+          queueState.subcribers.get(selector) match {
+            case None =>
+              (queueState, Some(ScalaQueue(queueState.last)))
+            case r @ Some(q) =>
+              if (q.isEmpty) (queueState, None)
+              else {
+                (queueState.copy(
+                   subcribers = queueState.subcribers + (selector -> ScalaQueue.empty)),
+                 r)
+              }
           }
 
-        def subscribers: Stream[F, Int] =
-          Stream
-            .bracket(Sync[F].delay(new Token))(token => pubSub.unsubscribe(Left(Some(token))))
-            .flatMap { token =>
-              Stream.repeatEval(pubSub.get(Left(Some(token)))).flatMap {
-                case Left(s)  => Stream.emit(s.subcribers.size)
-                case Right(_) => Stream.empty //impossible
-              }
-            }
+        def empty(queueState: State[A]): Boolean =
+          false
+
+        def subscribe(selector: (Token, Int), queueState: State[A]): (State[A], Boolean) =
+          (queueState.copy(
+             subcribers = queueState.subcribers + (selector -> ScalaQueue(queueState.last))),
+           true)
+
+        def unsubscribe(selector: (Token, Int), queueState: State[A]): State[A] =
+          queueState.copy(subcribers = queueState.subcribers - selector)
       }
-    }
+
+  }
 }
