--- experimental/src/main/scala/fs2/experimental/concurrent/pubsub/PubSubStrategy.scala
@@ -1,4 +1,0 @@
-package fs2.experimental.concurrent.pubsub
-
-trait PubSubStrategy[I, O, S, Selector]
-    extends fs2.concurrent.pubsub.PubSubStrategy[I, O, S, Selector]
